#148 Add Android Integration Tests and Refactor Testing
IntegrationTests.XamarinAndroid.csproj
- created as "Android Unit test" project
- changed Target Framework from 5.1 down to 4.4
- added reference to RealmNet.XamarinAndroid
- added reference to IntegrationTests.Shared

Interop.XamarionAndroid/AssemblyInfo.cs
- added [assembly: InternalsVisibleTo("IntegrationTests.XamarinAndroid")]

In IntegrationTests.shared project
- AbstractIntegrationTests.cs
  - renamed IntegrationTests.cs
  - made non-abstract
  - BaseSetup renaned Setup
  - removed empty virtual Setup
  - removed GetTempDatabasePath
  - set CoreProvider

- AbstractPerformanceTests
  - renamed PerformanceTests.cs
  - made non-abstract
  - BaseSetup renaned Setup
  - removed empty virtual Setup
  - removed GetTempDatabasePath
  - set CoreProvider
  - moved  RawPerformanceTest over from Win32PerformanceTest

- Win32PerformanceTest - gutted so just there to run tests
- Win32IntegrationTest - gutted
