+++ realm/Tests/IntegrationTests.XamarinAndroid/Properties/AndroidManifest.xml
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" android:versionCode="1" android:versionName="1.0" package="IntegrationTests.XamarinAndroid">
+	<uses-sdk />
+	<application android:label="IntegrationTests.XamarinAndroid">
+	</application>
+</manifest>
