--- realm/Tests/IntegrationTests.XamarinAndroid/TestsSample.cs
@@ -1,48 +1,0 @@
-﻿using System;
-using NUnit.Framework;
-
-namespace IntegrationTests.XamarinAndroid
-{
-    [TestFixture]
-    public class TestsSample
-    {
-        
-        [SetUp]
-        public void Setup ()
-        {
-        }
-
-        
-        [TearDown]
-        public void Tear ()
-        {
-        }
-
-        [Test]
-        public void Pass ()
-        {
-            Console.WriteLine ("test1");
-            Assert.True (true);
-        }
-
-        [Test]
-        public void Fail ()
-        {
-            Assert.False (true);
-        }
-
-        [Test]
-        [Ignore ("another time")]
-        public void Ignore ()
-        {
-            Assert.True (false);
-        }
-
-        [Test]
-        public void Inconclusive ()
-        {
-            Assert.Inconclusive ("Inconclusive");
-        }
-    }
-}
-
